Add files via upload
Starter.java is the main file that runs an instance of MenuGen.java. MenuGen.java enables the other classes to be utilized;

Operation.java is the Interface that the other classes use.
Uniqueness.java checks if the array passed to it is unique (No repeating values.)
LargestInteger.java finds the largest integer within an array passed to it.
GreatestSum.java adds three consecutive indexes of the array, and finds the greatest possible sum in each of the consecutive indexes.